Function mimty::ffi::mimty_file

pub extern fn mimty_file(path: *const u8) -> *const u8

Attempts to determine file MIME type by either examining the files extension, then by examining its content if extension identification was unsuccessful.

char* mimty_file(char *path);

Upon success, the MIME type of the file will be returned. Null is returned otherwise. This function takes and returns null-terminated strings compatible with C. Returned strings remain valid for the life of the library.

Note

At present, paths that contain invalid utf8 will cause this function to return null. This will be fixed as soon as from_osstr or as_osstr functionality is implemented for [u8]. If this is an issue, consider passing a dummy filename with a matching extension and using mimty_blob to identify the contents.