Function mimty::file

pub fn file<P: AsRef<Path>>(path: P) -> MimeType

Attempt to identify a file by its extension and contents

When identifying the MIME type of a file, proceedence is given to the extension of the file. If an extension is present, the library will not attempt to investigate further; i.e. the file does not need to exist for identification to succeed.

If an extension is not present, the library will access the file to identify it by its contents.

This function will return None if the file does not exist, the process does not have sufficiant permissions to read the file, or if the MIME type of the file could not be determined.